Milestone 3 - Role Choice
User can now pick a role. Each role file has attributes in it that are hard coded and returned to the App.py file through initializing the role. If a user picks an invalid number when picking a role, they are asked to pick again until they pick a valid number. All that is now set and I can now begin creating the actual gameplay.