Milestone 4.5
For my third game I am creating my own version of Blackjack. I have created the basic card generation where it returns a visual of your card, the rank and suit. As for blackjack rules, if you select an Ace you are able to pick if its value is 1 or 11